# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loudthotz Poetry Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Poem {
  id: number;
  title: string;
  author: string;
  country: string;
  content: string;
  publishedAt: string;
  averageRating: number;
  ratingCount: number;
  isFeatured?: boolean;
  /** @nullable */
  season?: string | null;
  /** @nullable */
  theme?: string | null;
}

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Submission {
  id: number;
  title: string;
  author: string;
  country: string;
  content: string;
  submittedAt: string;
  status: SubmissionStatus;
  /** @nullable */
  adminNote?: string | null;
}

export interface SubmissionInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  author: string;
  /** @minLength 1 */
  country: string;
  /** @minLength 10 */
  content: string;
}

export interface RatingInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  stars: number;
}

export interface LivestreamStatus {
  isLive: boolean;
  viewerCount: number;
  title: string;
  description: string;
  season: string;
  episode: number;
  /** @nullable */
  streamUrl: string | null;
  /** @nullable */
  embedUrl?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
}

export interface LivestreamSession {
  id: number;
  title: string;
  description: string;
  date: string;
  season: string;
  episode: number;
  /** @nullable */
  recordingUrl: string | null;
  theme: string;
}

export interface Book {
  id: number;
  title: string;
  subtitle: string;
  description: string;
  price: string;
  amazonUrl: string;
  accentColor: string;
  /** @nullable */
  coverTagline?: string | null;
}

export interface DonationInput {
  amount: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  message?: string | null;
}

export interface DonationResult {
  success: boolean;
  message: string;
  /** @nullable */
  reference?: string | null;
}

export interface SiteStats {
  totalPoems: number;
  totalPoets: number;
  totalCountries: number;
  totalSessions: number;
  upcomingEventTitle: string;
  /** @nullable */
  upcomingEventDate: string | null;
}

export type ListPoemsParams = {
search?: string;
sort?: ListPoemsSort;
page?: number;
limit?: number;
};

export type ListPoemsSort = typeof ListPoemsSort[keyof typeof ListPoemsSort];


export const ListPoemsSort = {
  latest: 'latest',
  popular: 'popular',
  alphabetical: 'alphabetical',
} as const;