# lib/api-zod/src/generated/types/donationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loudthotz Poetry Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface DonationInput {
  amount: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  message?: string | null;
}