# lib/api-zod/src/generated/types/donationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loudthotz Poetry Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface DonationResult {
  success: boolean;
  message: string;
  /** @nullable */
  reference?: string | null;
}