# lib/api-zod/src/generated/types/livestreamStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loudthotz Poetry Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface LivestreamStatus {
  isLive: boolean;
  viewerCount: number;
  title: string;
  description: string;
  season: string;
  episode: number;
  /** @nullable */
  streamUrl: string | null;
  /** @nullable */
  embedUrl?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
}