# lib/api-zod/src/generated/types/poem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loudthotz Poetry Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Poem {
  id: number;
  title: string;
  author: string;
  country: string;
  content: string;
  publishedAt: Date;
  averageRating: number;
  ratingCount: number;
  isFeatured?: boolean;
  /** @nullable */
  season?: string | null;
  /** @nullable */
  theme?: string | null;
}