# lib/api-zod/src/generated/types/ratingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loudthotz Poetry Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface RatingInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  stars: number;
}