# lib/api-zod/src/generated/types/siteStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loudthotz Poetry Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteStats {
  totalPoems: number;
  totalPoets: number;
  totalCountries: number;
  totalSessions: number;
  upcomingEventTitle: string;
  /** @nullable */
  upcomingEventDate: string | null;
}