# lib/api-zod/src/generated/types/submission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loudthotz Poetry Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmissionStatus } from './submissionStatus';

export interface Submission {
  id: number;
  title: string;
  author: string;
  country: string;
  content: string;
  submittedAt: Date;
  status: SubmissionStatus;
  /** @nullable */
  adminNote?: string | null;
}